Task5/llava-interleave-qwen-0.5b-hf/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu124
torch==2.4.0+cu124
torchvision==0.19.0+cu124
torchaudio==2.4.0+cu124
transformers==4.43.3
accelerate==0.33.0
sentencepiece
safetensors
huggingface-hub>=0.35.1
pillow>=10.0
